requirements.txt
streamlit
deepface
pillow
